Refactor runtime type handling
