+++ studio/src/components/RuntimeProvider/index.ts
@@ -1,1 +1,2 @@
 export * from "./RuntimeProvider";
+export * from "./types";
